refactor(binary): share post-download artifact helpers in binary_artifact
agent_binary and capsule_binary each carried their own copy of the
post-download artifact steps: container_arch, chmod_executable,
hash_file_sha256, and the tar.gz member extractor. Move one copy into a
new src/binary_artifact module and route both callers (and the
build-jackin-capsule bin) through it. net.rs keeps the network transfer;
binary_artifact owns everything that happens to the bytes once on disk.

- hash_file_sha256 is now a single sync fn; both callers wrap it in
  spawn_blocking at the call site (capsule already did, agent's internal
  spawn_blocking wrapper is gone).
- The two extractors unify into extract_tar_gz_member(archive, member,
  dest): basename match + is_file guard + entry.unpack. The agent path
  switches from File::create + io::copy to unpack; final file mode is
  unchanged because both callers chmod 0o755 immediately after.
- The hash + extract tests move from capsule_binary to binary_artifact.

Also reword a net.rs comment to name the real macro (debug_log!) instead
of the conceptual cdebug! tier name.

No behavior change. 2238 lib tests pass; clippy --all-targets clean.
